Setting up the Game 01
Unity 3D Roll-a-ball lesson 01
Created project
Add Ground
Add top and bottom light
Add ball

This lesson thought me how to add objects to the game and organise them.
It is possible to create empty objects and use them to nest related
objects together. This was illustrated with the Lighting object